fix: fix peformance issues with group/member tabs
basically we were filtering members by their groups (and vice versa) all at once while opening the tab. this is bad!

on top of that, the group list was doubly affected because we were looping through the entire group list AND member list... while we only needed to loop through the member list

in any case, the filtering is now postponed to whenever you open the group view tab for that individual member/group. should make things a lot speedier